version9.0.1 赛车 v6.1 fix

+++ src/main/java/cn/cherish/dubo/dubo/service/v6/CarResult.java
@@ -63,7 +63,7 @@ public class CarResult {
              */
 
             private String preDrawTime;
-            private int preDrawIssue;
+            private long preDrawIssue;
             private String preDrawCode;
             private int sumFS;
             private int sumBigSamll;
